% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model from data}
\usage{
fit_model(
  genotype,
  phenotype,
  model_type = "regularised",
  cores = 4,
  iter = 2500,
  adapt_delta = 0.95,
  max_treedepth = 15
)
}
\arguments{
\item{genotype}{Genotype dataframe}

\item{phenotype}{Phenotype vector}

\item{model_type}{Type of model used to fit. "baseline", "regularised" and "blosum_enhanced" are accepted.}

\item{cores}{STAN: Number of cores}

\item{iter}{STAN: Number of iterations}

\item{adapt_delta}{STAN: Adapt delta}

\item{max_treedepth}{STAN: Max tree depth}
}
\value{

}
\description{
Fit model from data
}
\examples{
fit_model(X, y, model_type="regularised")
}
